Large scale refactoring of classes and command structure to allow for better code readability - P
