+++ src/main/java/com/paralli/phantom_waypoints/models/Player_waypoint.java
@@ -1,4 +1,4 @@
-package com.paralli.phantom_waypoints.classes;
+package com.paralli.phantom_waypoints.models;
 
 import java.io.Serializable;
 import java.util.List;
